GLImageNativePixmap: specify YUV color space.
This CL updates GLImageNativePixmap::Initialize() to specify the
YUV-to-RGB conversion matrix. Until now, judging by the fact that our
video.DrawOnCanvas.* tests have been passing, it looks like we've been
assuming EGL_ITU_REC601_EXT. However, a recent driver uprev revealed
that we cannot rely on this matrix being the default in the driver, so
Chrome needs to specify it.

Using EGL_ITU_REC601_EXT always is not correct, so this is a workaround
to go back to a known state. A longer-term solution (tracked in
b/220336463) involves plumbing enough information to GLImageNativePixmap
so that we can determine the correct value.

Bug: b:219444479, b:220336463
Test: tast run ${jacuzzi} video.DrawOnCanvas.*
Change-Id: I1771957c6fedaf6db58bbf64f1e5f77c01759fc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3473918
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974358}