Implement the Override List for Page Topics Annotations
Looks for a gzipped, binary protobuf file that contains a lookup table
for PageTopics model executions, short-cutting the model execution
entirely if the given domain is in the lookup table.

The override list is loaded when the model file would also expect to be
loaded, and unloaded at the same time.

See bug for design doc.

Bug: 1295766
Change-Id: I1ffc19c0f86be8993f3e692ad4383cb71bbdeb6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469662
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974400}