Add API to observe when ambient animation photos get refreshed.
This API will be used in the next CL. Whenever the photos in the
animation get refreshed, the attribution text in the animation also
needs to get refreshed. Thus, the observer method will be used
as a trigger for a different class that handles setting attribution
in the animation.

Bug: b:217271404

Change-Id: Ia21e8a22c1070e97f13f05a590a78e7d3250d0b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3485587
Reviewed-by: Tao Wu <wutao@chromium.org>
Commit-Queue: Eric Sum <esum@google.com>
Cr-Commit-Position: refs/heads/main@{#974408}