Modify the non mojom function check condition in AppUpdate.
AppUpdate is created by 2 scenarios:
1. ForAllApp, ForApp, ForEachApp, ForOneApp in AppRegistryCache
2. OnAppUpdate callback in AppRegistryCache

For non mojom call, state_, delta_ are set in AppUpdate by ForAllApp,
ForApp and OnAppUpdate non mojom call.
For mojom call, mojom_state_, mojom_delta_ are set by ForEachApp,
ForOneApp and OnAppUpdate mojom call.

So in AppUpdate, we can check state_, delta_ to decide using mojom or
non mojom App struct, and not depend on the flag. So that we can call
ForAllApp, ForApp to access the non mojom App struct without depending
on flag.

This CL modifies the checking condition in AppUpdate to remove the
dependency on the flag, and modify tests to verify ForAllApp, ForApp
with flag on and off.

BUG=1253250

Change-Id: I5f70fd0e927e21a03574e71d963777c38b3862ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3482116
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974409}