Add crash keys for empty URL restore with multiple entries
We expected that restoring a NavigationEntry with an empty URL is only
possible when we're restoring the initial NavigationEntry, which should
be the only entry in the session history, which is asserted via a
CHECK for the session history size. Some reports of this CHECK failing
was seen in the wild, so this CL adds crash keys (and changes the
CHECK to a DumpWithoutCrashing) to investigate.

Bug: 1299335
Change-Id: Id0f18034826a875dc7321d5379045517455b4d41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3474284
Auto-Submit: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974429}