Post data URL bitmap fetcher callback on correct thread
Bitmap fetcher forces async responses when the result can be calculated
synchronously (for data URLs). This fixes the thread on which the
response is posted - the task should be posted to the current sequenced
task runner, and no thread pool. I assume this was result of unintuitive
API for posting tasks to thread pool (which was named so it was easily
mistaken for API that posts to the current thread).

BUG=1300329

Change-Id: I38b06885afebba5c9f48239b0a2e3dcf768ff5c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3485002
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974430}