Make ProjectorClientImpl implement ProjectorAnnotatorController
We assumed the ProjectorAnnotatorController to be a singleton, but the
AnnotatorMessageHandler (its implementation) is not a singleton.
The fix is to have the ProjectorClientImpl (a singleton) implement the
ProjectorAnnotatorController.
In addition, the creation of the AnnotatorMessageHandler is being moved
to the TrustedProjectorAnnotatorUI. Multiple WebUIs were being created
for one projector recording, so the AnnotatorMessageHandler was being
destroyed when the first UI it was embedded in was destroyed.
Now, we create a new handler every time a new WebUI is created
(through TrustedProjectorAnnotatorUI), and register the new
handler in ProjectorClientImpl.

b: 216858461
Change-Id: I610f069e83982f9043582af38871abdc5bd01b20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3456900
Reviewed-by: Li Lin <llin@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Courtney Wong <courtneywong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974431}