Add a max report limit to SCT auditing hashdance
This adds a new local state pref in the browser to track the number of
SCT auditing hash dance reports sent, so a "max reports" limit can be
applied per client. This is implemented as NetworkContextClient APIs
for the SCT auditing code in the Network Service to query and
increment the report count.

Bug: 1249925
Change-Id: Ibb30630627bc03fd2d5c629f5ffe6641fd57c811
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3457062
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974444}