Use String IDs for Tutorial Body/Title text, fix windows display bug
body/title_text is switched for body/title_text_id, and then
the tab group tutorial is changed to use localized string ids.

there was also an issue where on windows 10 and chromeOS the
TabStripRegionView would not set the ElementID property because of early
returns. in order to fix this the SetProperty call is moved up above
the OS specific calls.

Bug: 1299864
Change-Id: I728c466f47f2a71fe9c16219aea81d5a315cc01b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3469821
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974451}