[Extensions] Make notifications custom bindings use failure callback
Previously in the custom bindings for chrome.notifications.create and
chrome.notifications.update if image loading failed for some reason we
would call the callback with lastError set as well as passing the
callback an argument (false for 'update' and the ID for 'create').
However failures in APIs are not meant to return arguments and this
pattern is incompatible when using promises for asynchronous calls.
The values were also not actually useful, as for 'create' the ID would
just be passing back the ID you called the API with and for 'update' it
would always return false in the case of this error.

This CL removes the argument returned on failure and uses the
failureCallback recently added to HandleRequest hooks to handle
rejecting the associated promise or setting lastError and calling the
callback. Although this is a slight behavior change, it was a very
niche case that was both undocumented and not covered in tests.

Bug: 1220910
Change-Id: Ifd960b938d64b589cbdd33bc089e1abd9ae544a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3473887
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974453}