[Files SWA]: Better identify various volume not found causes.
In number of places we log the same message "Volume not found". Without
access to C++ logs we do not get information what function causes this
and what volume could not be found. This CL adds a string identifying
the source function and the ID of the volume.

Bug: 1299920
Change-Id: I8d2dddb28d48a555bffd57c6b07d2fdf33ceb675
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3486237
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Bo Majewski <majewski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974485}