[CrOS Settings] Move authentication type out of VPN types
Currently the VPN types list on the settings pages contain the
combinations of VPN types and authentication types for L2TP/IPsec. Since
we will add IKEv2 VPN in the future, which will have 3 kinds of
authentication, the list will be very long if we keep this method.
This patch moves the authentication type out of VPN types and makes it
as a separate dropdown menu, and its default value will be "PSK" since
PSK authentication is the majority of the usage according to the
metrics. Also add a unit test to verify if the items in the UI is shown
or hidden properly.

  in the UI.

Bug: b/214880538
Test: unit tests; manally checked that able to create a L2TP/IPsec VPN
Change-Id: I411f7183ed510dc71c49ee86217fd60041da6751
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3478936
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974517}