android: histogram for SafeBrowsing time at first navigation
Add a histogram to record the time spent in the SafeBrowsing backend
(GmsCore) for the first navigation since startup. Exclude CCT and record
it at the same time and iff the .Tabbed commit histogram is recorded.
This should make it easier to account for any missing samples.

The SB response result makes a hop from the handler thread, to IO and
then to the UI thread. To avoid carrying the time delta along with it,
use atomicops to read/update the value.

May add a test at a later commit we find a good way to mock out the
SafeBrowsingApiBridge in an instrumentation test. Please advise.

The checkdeps complained, and was silenced by copying a rule that is
already present in android_webview/java/DEPS and chrome/browser/DEPS.
Hopefully this is OK?

Bug: 1296097
Change-Id: Ic118f573322d07ed108f50045b17bb27c9f08605
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3477161
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Hao Dong <spdonghao@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974677}