[Tab Management] Flaky fix by replacing isDisplayed
This test is flaking, but never outright failing, on Android M x86.
See bug for examples.

I cannot reproduce locally, but it appears that the tab switcher button
is sometimes not qualifying as "isDisplayed()", but it is visible and
clickable. As such, replace "isDisplayed()" checks for effective
visibility. This is acceptable as the point of this test is to test a
dialog is shown via the expected flow not that the objects required to
reach that dialog are displayed correctly.

Bug: 1300346
Change-Id: Ic7f6015772c7754ffab8bbb988c010ba4d480910
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3488666
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974702}