CronetURLRequestContext: network to URLRequestContext mapping
This change adds the possibility for CronetURLRequestContext to hold
multiple URLRequestContexts. In the CL following this, it will be
renamed to CronetContext, to reflect the fact that this is no longer a
wrapper around a single URLRequestContext.

One of these contexts is considered to be the default one and will be
used whenever a request does not specify a target network. Additional
contexts are instead bound to a specific network and will be used when
a request specifies the network they are bound to.

The default context is constructed at init time and its lifetime is
directly linked to its CronetURLRequestContext's lifetime.

Additional contexts are instead created on the fly, whenever a requests
targeting `network` is created and a context for `network` does not
already exists.
Additionally, they are automatically destroyed once `network`
disconnects. To do this now CronetURLRequestContext::NetworkTasks now
can also listen to NetworkChangeNotifier::NetworkObserver type of
events.

Note: The purpose of this CL is to make this change transparent to
existing usage of CronetURLRequestContext. As such,
CronetURLRequestContext::GetURLRequestContext additional parameter is
optional and default to kInvalidNetworkHandle, which is used to
represent the default context. Tests will be added in a following CL,
when the optional parameter will start being used.

Bug: 1284972
Change-Id: I0a57109fa9da3ad1b60b0231d9baf905a959e6f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3452308
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Stefano Duo <stefanoduo@google.com>
Cr-Commit-Position: refs/heads/main@{#974720}