[PartitionAlloc] Move the spinning part of SpinningMutex out of line
This is mostly relevant now that this lock has more callers. It is also
meant to make the "happy path" of the non-thread cache case faster. This
is still quite hot code, as the thread cache is only enabled for a
single partition.

Namely,
- All the experiments leading to split partitions (e.g. aligned, BRP,
  *Scan splits) have multiple partitions
- In renderers, the buffer partition is hot. It is in particular used to
  back WTF objects, including strings.

On Linux x86_64, this CL shortens PartitionFree(), see the last
instruction offsets below (from +1317 to +1250):

BEFORE:
   0x00000000011bb93d <+1245>:	lock and %rax,(%rbx,%rcx,1)
   0x00000000011bcc65 <+1317>:	jmp    0x11bc7a5 <PartitionFree()+101>
End of assembler dump.

AFTER:
   0x00000000011bcc60 <+1312>:	lock and %rax,(%rbx,%rcx,1)
   0x00000000011bb942 <+1250>:	jmp    0x11bb4c9 <PartitionFree()+105>
End of assembler dump.

It also moves the return of the "relatively happy path" (i.e., locking,
but no other complications) from offset +774 to +719, see below:

BEFORE:
   0x00000000011bca46 <+774>:	pop    %rbp
   0x00000000011bca47 <+775>:	jmp    0x11c5160 <_ZN15partition_alloc8internal13SpinningMutex9FutexWakeEv>

AFTER:
   0x00000000011bb72e <+718>:	pop    %rbp
   0x00000000011bb72f <+719>:	jmp    0x11c3c10 <_ZN15partition_alloc8internal13SpinningMutex9FutexWakeEv>

(This is a tail call to the "futex wake" syscall wrapper, which is not
the happiest path, since it happens when the mutex is contended, but is
easier to visualize).

Finally, through a happy accident, it also shortens the fastest path
slightly (with a thread cache), with a return moved from offset +441 to
+429:

BEFORE:
   0x00000000011bc8f8 <+440>:	pop    %rbp
   0x00000000011bc8f9 <+441>:	ret

AFTER:
   0x00000000011bb60c <+428>:	pop    %rbp
   0x00000000011bb60d <+429>:	ret
(caused by different stack spilling, but this is pretty brittle)

The main reason for the shorter function is the change from:
   0x00000000011bc964 <+548>:	lock cmpxchg %ecx,(%r15)
   0x00000000011bc969 <+553>:	je     0x11bc998 <PartitionFree()+600>
   0x00000000011bc96b <+555>:	mov    %rdx,%r14
   0x00000000011bc96e <+558>:	mov    %edi,%eax
   0x00000000011bc970 <+560>:	test   %edi,%edi
   0x00000000011bc972 <+562>:	jle    0x11bc97d <PartitionFree()+573>
   0x00000000011bc974 <+564>:	pause
   0x00000000011bc976 <+566>:	add    $0xffffffff,%eax
   0x00000000011bc979 <+569>:	jne    0x11bc974 <PartitionFree()+564>
   0x00000000011bc97b <+571>:	add    %edi,%ebx
   0x00000000011bc97d <+573>:	add    %edi,%edi
   0x00000000011bc97f <+575>:	cmp    $0x10,%edi
   0x00000000011bc982 <+578>:	cmovge %esi,%edi
   0x00000000011bc985 <+581>:	cmp    $0x40,%ebx
   0x00000000011bc988 <+584>:	mov    %r14,%rdx
   0x00000000011bc98b <+587>:	jl     0x11bc95b <PartitionFree()+539>
   0x00000000011bc98d <+589>:	mov    %r15,%rdi
   0x00000000011bc990 <+592>:	call   0x11c51c0 <LockSlow()>

to:
   0x00000000011bb66d <+525>:	lock cmpxchg %ecx,(%r15)
   0x00000000011bb672 <+530>:	je     0x11bb680 <PartitionFree()+544>
   0x00000000011bb674 <+532>:	mov    %r15,%rdi
   0x00000000011bb677 <+535>:	call   0x11c3b50 <AcquireSpinThenBlock()>

Which is the outlining of the spinning part.

Bug: 998048
Change-Id: Iee83d30989d153360471c1b729ace2c9bb35e191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3484760
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974726}