Add test coverage for syncer::StringOrdinal
This is a preparation to immediately merge Ordinal with StringOrdinal
in crrev.com/c/3488388. The goal is to migrate the unit tests while
also making sure that behavior gets preserved by the other CL. Since
Ordinal's behavior is weird (see crbug.com/1300525#c1), we do the best
we can, i.e. replace N in the general case with the N-th letter in the
0-indexed alphabet. Some of the tests for LongOrdinal and LargeOrdinal
are also removed since they are no longer interesting.

Bug: 1300525
Change-Id: I3f1d06b0a0ce42804827a771aa4c78af97972fb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3488271
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#974778}