[Extension bindings] Avoid crashing in CompleteHandleRequestHelper
This CL ensures that we fail gracefully if no associated active request
was found for a request ID when we are resolving a handle request hook
in the JS custom bindings for an extension API.

In theory this should never happen because we will have always stored
the data for an active request if this function is being called and the
handle request hook should only ever be calling the resolution callbacks
(that trigger this function) a single time. However if a handle request
hook erroneously calls the resolution callbacks more than once, the
stored request for that request ID will have already been erased and
find() will come back with the end iterator. Previously this could
cause a crash when trying to call erase() with the end iterator, so
now instead we bail early.

Bug: 1298409
Change-Id: Id847d5478d723b6139dfe4e7a1d5b2ba2e23e793
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3472219
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974791}