Fix glitch where the user deletes a tab while typing in the Omnibox.
This change detects the scenario where the user typed a URL of another
tab in the omnibox (triggering the Switch-to-Tab suggestion) and
deleted the tab right before asking to switch to it.

When this happens, both Java and C++ Tab objects continue to live, but
the TabModel no longer has a tab to jump to.

Bug: 1300447
Change-Id: Ic1b5d34a27536ef5baa5b77d9fda85b16476fac7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3488576
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#974798}