Add support for Android mixed-C++/Rust unit tests.
//testing:

We add a mixed_shared_library GN target which is used to build the
unit test library, and placed into the Android APK. This allows the
shared library to include Rust code. And then we enable the
rust_gtest_interop_unittests on Android.

In doing so, we get a Rust warning due to a double-underscore in a GN
target name, which is translated into the crate name:

[5195/5234] RUST obj/testing/rust_gtest_interop/librust_gtest_interop_unittests__library_rs.rlib
warning: crate `rust_gtest_interop_unittests__library_rs` should have a snake case name
  |
  = note: `#[warn(non_snake_case)]` on by default
  = help: convert the identifier to snake case: `rust_gtest_interop_unittests_library_rs`

We can fix this by removing the double-underscore from the crate name,
without changing the actual .so file name, since many things look for
the .so name. The crate name is only used for the purpose of building
the test library crate, which has no dependents. And they are able to
choose their own name for it anyhow.

//build/rust:

When we enable the rust_gtest_interop_unittests on Android, it exposes
some cross-toolchain issues for proc-macros.

1) If the proc-macro target is emitted for the $current_toolchain
on Android ARM, the GN generation fails, as there's no rust_macro
tool there, because `rustc_can_link` is false. What this points
out is that we should not be emitting the target for the
$current_toolchain at all, since it is only used on the host,
in the $host_toolchain.

2) When the proc-macro target is depended on, it can jump to the
appropriate toolchain as demonstrated by fuchsia:
 https://fuchsia.googlesource.com/fuchsia/+/refs/heads/main/build/rust/rustc_macro.gni#258

So when we don't emit the proc-macro target, we instead emit a
group that points to the correct toolchain.

3) In the process we have to mark a bunch of variables as
not_needed when the proc-macro target is not emitted. This
pointed out that the `crate_root` variable in rust_target() is
redundant, since we set the `crate_root` in the actual target
rule from `_crate_root`. So we remove it.

4) The proc-macro is built on the host toolchain but it has to
emit rust code for the target toolchain. This is a problem because
the host toolchain sees c_char as i8, but the code it has to emit
will see c_char as u8. So we have to explicitly cast the char
constants that the macro wrote out to c_char.

R=adetaylor@chromium.org, johnchen@chromium.org

Bug: 1293979, 1299883, 1297592
Change-Id: Ief58f96d12f386a69622701263f8ffe1ed0d4dd2
Cq-Include-Trybots: luci.chromium.try:linux-rust-x64-rel,android-rust-arm-rel
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3485799
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Adrian Taylor <adetaylor@google.com>
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974821}