[Fast Pair] Fix device images not showing when device ID changes
Adds in a call to fetch device images after a device has connected,
which will save a mapping from the current device ID to model ID. This
is necessary since the original call to fetch device images happens
earlier with a potentially different device ID. Also add in more tests
and logs for DeviceIdMap to make debugging this issue easier in the
future.

change fixes the regression.

Change-Id: I4a3439eb34798d9183b6e19106a18d696c714a2a
Bug: b:220913514
Test: Tested manually by pairing JBL Tune 125TWS, confirmed that this
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3482111
Reviewed-by: Shane Fitzpatrick <shanefitz@google.com>
Commit-Queue: Daniel Classon <dclasson@google.com>
Cr-Commit-Position: refs/heads/main@{#974836}