[infra] Use chromium as project in builder configs.
Currently, when generating the builder configs on a branch project (e.g.
chromium-m100), the branch project's name is used in the generated
builder IDs. The project in the builder IDs determines the luci project
value for the build's BuilderSpec in the recipe, which is used when the
builder is triggered by another builder. This requires re-generating the
configs when cherry-picking so the Rubber Stamper can't be used.

Builders that have their config recipe-side already manage this
situation since a single version of the config exists for all branched
versions of a builder. This is handled by the project_trigger_overrides
field in the chromium_tests property. This fields overrides the project
when used for triggering so that the correct version of the builder will
be triggered.

This change generates the chromium project regardless of the branch and
relies on the project_trigger_overrides value that is already set on
branch projects to trigger the correct branched version of builders and
will allow for clean cherry-picks that don't require re-generating the
configs.

Bug: 868153
Change-Id: I639cde2ecf2aba7d6dd54aac2e316cba3bdadd41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3487971
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Erik Staab <estaab@chromium.org>
Commit-Queue: Erik Staab <estaab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#974846}